+++ HISTORY.md
@@ -1,5 +1,9 @@
 ## Changelog ##
 
+### In Development (Dev time: 1 hours) ###
+- New Features
+	- Add actions that are executed before and after a query is executed.
+
 ### 2015-07-21 - v5.5 (Dev time: 50 hours) ###
 - New Features
     - Added the ability to delete users based on user meta. (Issue #79)