+++ HISTORY.md
@@ -1,6 +1,6 @@
 ## Changelog ##
 
-### In Development (Dev time: 2.0 hours) ###
+### In Development (Dev time: 2.5 hours) ###
 - New Features
 	- Add actions that are executed before and after a query is executed.
 	- Add actions that are executed before and after scripts and styles are enqueued.
@@ -8,6 +8,7 @@
 - Enhancements
 	- Add compatibility with "The Events Calendar" plugin.
 	- Add compatibility with "WooCommerce" plugin. (Issue #111)
+	- Display warning in the System Info page if certain required options are disabled. (Issue #106)
 
 ### 2015-07-21 - v5.5 (Dev time: 50 hours) ###
 - New Features